Updated the Recipe Details UI and Fixed Folder structure
